finish

+++ src/main/java/com/security/jwt/services/impl/LibroServiceImpl.java
@@ -6,6 +6,7 @@ import java.util.Optional;
 import org.springframework.stereotype.Service;
 
 import com.security.jwt.repositories.LibroRepository;
+import com.security.jwt.repositories.PrestamoRepository;
 import com.security.jwt.repositories.entities.LibroEntity;
 import com.security.jwt.services.LibroService;
 
@@ -16,7 +17,9 @@ import lombok.AllArgsConstructor;
 public class LibroServiceImpl implements LibroService {
 
     private LibroRepository libroRepository;
+    private PrestamoRepository prestamoRepository;
 
+    @SuppressWarnings("null")
     @Override
     public LibroEntity save(LibroEntity libro) {
         return libroRepository.save(libro);
@@ -27,10 +30,11 @@ public class LibroServiceImpl implements LibroService {
         return (List<LibroEntity>) libroRepository.findAll();
     }
 
+    @SuppressWarnings("null")
     @Override
     public boolean edit(Long id, LibroEntity libro) {
         Optional<LibroEntity> editLibro = libroRepository.findById(id);
-        if (editLibro.isPresent()) {
+        if (editLibro.isPresent() && libro.getInventario() > 0) {
             editLibro.get().setTitulo(libro.getTitulo());
             editLibro.get().setAutor(libro.getAutor());
             editLibro.get().setGenero(libro.getGenero());
@@ -42,12 +46,15 @@ public class LibroServiceImpl implements LibroService {
         return false;
     }
 
+    @SuppressWarnings("null")
     @Override
     public boolean deleteById(Long id) {
         Optional<LibroEntity> deleteLibro = libroRepository.findById(id);
         if (deleteLibro.isPresent()) {
-            libroRepository.delete(deleteLibro.get());
-            return true;
+            if (!prestamoRepository.existsByLibro(deleteLibro.get())) {
+                libroRepository.delete(deleteLibro.get());
+                return true;
+            }
         }
         return false;
     }
